+++ packages/core/src/tracing/instrumentation/loggers/pino.js
@@ -2,6 +2,7 @@
 
 'use strict';
 
+const { inspect } = require('util');
 const shimmer = require('shimmer');
 
 const requireHook = require('../../../util/requireHook');
@@ -40,14 +41,30 @@ function shimGenLog(originalGenLog) {
               const span = cls.startSpan('log.pino', constants.EXIT);
               span.stack = tracingUtil.getStackTrace(log);
               if (typeof mergingObject === 'string') {
+                // calls like logger.error('only a message')
                 message = mergingObject;
               } else if (mergingObject && typeof mergingObject.message === 'string' && typeof message === 'string') {
+                // calls like
+                // logger.error({ message: 'a message in the merging object'}, 'an additional  message as a string')
                 message = `${mergingObject.message} -- ${message}`;
               } else if (mergingObject && typeof mergingObject.message === 'string') {
+                // calls like
+                // logger.error({ message: 'a message in the merging object'}) or
+                // logger.error({ message: 'a message in the merging object: %s'}, /* non-string interpolation param */)
                 message = mergingObject.message;
-              } else if (typeof message !== 'string') {
-                message =
-                  'Log call without message. The Pino mergingObject argument will not be serialized by Instana for performance reasons.';
+              } else if (typeof message === 'string') {
+                // calls like
+                // logger.error({ /* merging object without message attribute */ }, 'a string message)
+                // Nothing to do, just use the given message (second argument) and ignore the first argument, which
+                // apparently has no message attribute
+              } else if (mergingObject != null) {
+                // Fallback for calls with an unknown shape, like:
+                // logger.error({ /* merging object without message attribute */ })
+                // Serialize the first argument, but only the first level, and also shorten it:
+                message = inspect(mergingObject, { depth: 1 }).substring(0, 500);
+              } else {
+                // If it is neither of those call patterns, we give up and do not capture a message.
+                message = 'Pino log call without message and mergingObject.';
               }
               span.data.log = {
                 message